+++ cli/Sources/TuistServer/Services/UploadBuildService.swift
@@ -35,13 +35,16 @@ enum UploadBuildServiceError: LocalizedError {
 public struct UploadBuildService: UploadBuildServicing {
     private let fullHandleService: FullHandleServicing
     private let multipartUploadArtifactService: MultipartUploadArtifactServicing
+    private let retryProvider: RetryProviding
 
     public init(
         fullHandleService: FullHandleServicing = FullHandleService(),
-        multipartUploadArtifactService: MultipartUploadArtifactServicing = MultipartUploadArtifactService()
+        multipartUploadArtifactService: MultipartUploadArtifactServicing = MultipartUploadArtifactService(),
+        retryProvider: RetryProviding = RetryProvider()
     ) {
         self.fullHandleService = fullHandleService
         self.multipartUploadArtifactService = multipartUploadArtifactService
+        self.retryProvider = retryProvider
     }
 
     public func uploadBuild(
@@ -55,12 +58,14 @@ public struct UploadBuildService: UploadBuildServicing {
         let accountHandle = handles.accountHandle
         let projectHandle = handles.projectHandle
 
-        let uploadId = try await startMultipartUpload(
-            client: client,
-            accountHandle: accountHandle,
-            projectHandle: projectHandle,
-            buildId: buildId
-        )
+        let uploadId = try await retryProvider.runWithRetries { [self] in
+            try await startMultipartUpload(
+                client: client,
+                accountHandle: accountHandle,
+                projectHandle: projectHandle,
+                buildId: buildId
+            )
+        }
 
         let parts = try await multipartUploadArtifactService.multipartUploadArtifact(
             artifactPath: filePath,
@@ -77,14 +82,16 @@ public struct UploadBuildService: UploadBuildServicing {
             updateProgress: { _ in }
         )
 
-        try await completeMultipartUpload(
-            client: client,
-            accountHandle: accountHandle,
-            projectHandle: projectHandle,
-            buildId: buildId,
-            uploadId: uploadId,
-            parts: parts
-        )
+        try await retryProvider.runWithRetries { [self] in
+            try await completeMultipartUpload(
+                client: client,
+                accountHandle: accountHandle,
+                projectHandle: projectHandle,
+                buildId: buildId,
+                uploadId: uploadId,
+                parts: parts
+            )
+        }
     }
 
     private func startMultipartUpload(