+++ cli/CHANGELOG.md
@@ -1,18 +1,30 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-## What's Changed in 4.176.0<!-- RELEASE NOTES START -->
+## What's Changed in 4.176.2<!-- RELEASE NOTES START -->
+
+### 🐛 Bug Fixes
+
+* mkdir data race by [@fortmarek](https://github.com/fortmarek) in [#10211](https://github.com/tuist/tuist/pull/10211)
+* write empty shard matrix output when selective testing skips all tests by [@fortmarek](https://github.com/fortmarek) in [#10205](https://github.com/tuist/tuist/pull/10205)
+
+
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.176.1...4.176.2
+
+## What's Changed in 4.176.1<!-- RELEASE NOTES START -->
 
 ### ⛰️  Features
 
 * add workspace-level DerivedData location support by [@davidpasztor](https://github.com/davidpasztor) in [#9693](https://github.com/tuist/tuist/pull/9693)
 ### 🐛 Bug Fixes
 
+* add retry logic to build uploads by [@fortmarek](https://github.com/fortmarek) in [#10210](https://github.com/tuist/tuist/pull/10210)
 * add nonisolated(unsafe) to generated plist accessors with Any type by [@fortmarek](https://github.com/fortmarek) in [#10195](https://github.com/tuist/tuist/pull/10195)
 
 
 
-**Full Changelog**: https://github.com/tuist/tuist/compare/4.175.0...4.176.0
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.175.0...4.176.1
 
 ## What's Changed in 4.175.0<!-- RELEASE NOTES START -->